uploaded newest GUI files
